Update ProfileBackup.ps1
Updated the form to support browse to directory.
Also fixed an error with the excel saving, now you are prompted to save the form as the file name wasn't working.
Figured out a way to store the variable in the textbox until I needed it when the other functions call the variable.